Update 0001-cnf-conformance-definition-proposal-process.md
First pass here before looking at the template itself. I really think that a best practice should lay out to type of workload or container it applies to. OSS/BSS, NF and infra containers may have best practices applied to them in unequal measure depending on their desired outcome. A container specifically designed to provision infra will need privilege for example, while a container hosting BGP would not. 

They alternative to this approach is for the definition of a CNF to be formalized (producing this is listed as a group outcome in our charter) and have its scope limited. From there we clarify that the best practices specifically apply to these "types" of workloads/containers.