Making goals more concise + removing conformance language
Removing old language related to conformance program. Groups main goal is identifying best practices for CNFs. The purpose of that is to get the benefits into developer and operator hands. Adherence to the principles is only a means to the end of gaining the benefits and not the purpose itself.

As part of showing best practices we will be finding practices that are specifically applicable to networking applications and showing how they can leverage the capabilities in K8s (as a platform and framework).