Flow correction to context
I had some discussions with @taylor on fixing the flow of how and when to provide the context of the workload. The discussion led us to believe that after the initial summaries and proposals to keep the pitch uninterrupted was important. The context coming before the user stories also seemed beneficial to help guide how the stories are drafted.

Last piece was to move the last sentence to the caveats section. It just feels more appropriate there given its content. A follow up PR for the template itself will be put in now that context has been accepted as a part of the overall process.