Removing defintions to get first merge through.
I'm going to pull out the definitions for Cloud Native and CNFs. I started a discussion at the beginning of this group's formation around defining CNF, and there was vocal opposition to this, many stating the definition existed already. As can be seen from this PR, this matter is not broadly agreed upon yet. As such, I think it would be better to take definitions on individually, so they can be debated/agreed upon without holding up PRs indefinitely.