Add in representation quota management procedures
We have previously specified that chairs should be seen to be
independent of each other and that tech leads should not end up
all from the same company.  This spells out the rules, and what to
do when we do find we are getting into trouble with an unbalanced
team.