reword test plan (#184)
* reword test plan

Since we are avoiding compliance, and not fully fleshing out a test plan that provides the "how" of "what" should be tested. I think we should consider changing the title section to "Testing Considerations"

* Update NNNN-cbpp-template.md

Changing to "objectives" based on @iawells suggestion. I agree that this word more accurately captures the intent.